+++ lib/ecto_db_scanner/result/check_constraint.ex
@@ -1,0 +1,8 @@
+defmodule EctoDBScanner.Result.CheckConstraint do
+  defstruct [:name, :expression]
+
+  @type t :: %__MODULE__{
+          name: String.t(),
+          expression: String.t()
+        }
+end