+++ lib/ecto_db_scanner/result/database.ex
@@ -1,7 +1,8 @@
 defmodule EctoDBScanner.Result.Database do
-  defstruct schemas: []
+  defstruct [:size_bytes, schemas: []]
 
   @type t :: %__MODULE__{
+          size_bytes: integer(),
           schemas: [EctoDBScanner.Result.Schema.t()]
         }
 end