+++ lib/ecto_db_scanner/result/index.ex
@@ -1,0 +1,15 @@
+defmodule EctoDBScanner.Result.Index do
+  defstruct [
+    :name,
+    :type,
+    :unique,
+    columns: []
+  ]
+
+  @type t :: %__MODULE__{
+          name: String.t(),
+          type: String.t(),
+          unique: boolean(),
+          columns: [String.t()]
+        }
+end