+++ lib/ecto_db_scanner/result/schema.ex
@@ -1,8 +1,9 @@
 defmodule EctoDBScanner.Result.Schema do
-  defstruct [:name, tables: []]
+  defstruct [:name, tables: [], sequences: []]
 
   @type t :: %__MODULE__{
           name: String.t(),
-          tables: [EctoDBScanner.Result.Table.t()]
+          tables: [EctoDBScanner.Result.Table.t()],
+          sequences: [EctoDBScanner.Result.Sequence.t()]
         }
 end