+++ lib/ecto_db_scanner/result/unique_constraint.ex
@@ -1,0 +1,8 @@
+defmodule EctoDBScanner.Result.UniqueConstraint do
+  defstruct [:name, columns: []]
+
+  @type t :: %__MODULE__{
+          name: String.t(),
+          columns: [String.t()]
+        }
+end