+++ lib/ecto_db_scanner/scanner.ex
@@ -19,6 +19,18 @@ defmodule EctoDBScanner.Scanner do
     argument :repo, input(:repo)
   end
 
+  step :sizes, EctoDBScanner.Steps.QuerySizes do
+    argument :repo, input(:repo)
+  end
+
+  step :indexes, EctoDBScanner.Steps.QueryIndexes do
+    argument :repo, input(:repo)
+  end
+
+  step :sequences, EctoDBScanner.Steps.QuerySequences do
+    argument :repo, input(:repo)
+  end
+
   step :detect_enums, EctoDBScanner.Steps.DetectEnums do
     argument :repo, input(:repo)
     argument :tables, result(:tables)
@@ -31,6 +43,9 @@ defmodule EctoDBScanner.Scanner do
     argument :columns, result(:columns)
     argument :enum_info, result(:detect_enums)
     argument :constraints, result(:constraints)
+    argument :sizes, result(:sizes)
+    argument :indexes, result(:indexes)
+    argument :sequences, result(:sequences)
   end
 
   return :assemble